Project restructuring and creation of empty app page files
