+++ app/(tabs)/events.tsx
@@ -1,4 +1,4 @@
-import { EventData, fetchEventData } from "@/app/services/eventsService";
+import { EventData, fetchEventData } from "@/services/eventsService";
 import { Component, useEffect, useState } from "react";
 import { Text, View, StyleSheet, ScrollView, ImageBackground } from "react-native";
 
@@ -92,7 +92,7 @@ const EventsPage = () => {
 <ScrollView contentContainerStyle={styles.scrollContainer}>
       <Text>DEMO Upcoming Events Page</Text>
       <ImageBackground
-        source={require("../../../assets/images/stockphoto.jpg")}
+        source={require("../../assets/images/stockphoto.jpg")}
         style={styles.frontimage}
         imageStyle={{ borderRadius: 10 }}
       >
@@ -106,7 +106,7 @@ const EventsPage = () => {
         <View key={index} style={styles.eventWrapper}>
           <View style={styles.eventContainer}>
             <ImageBackground
-              source={require('../../../assets/images/testbackground.jpg')} //needs to be replaced with actual images from Blob
+              source={require('../../assets/images/testbackground.jpg')} //needs to be replaced with actual images from Blob
               style={styles.image}
               imageStyle={{ borderRadius: 10 }}
             >
