+++ components/home/dynamic-event.tsx
@@ -1,8 +1,8 @@
-import { DynamicButtonProps, GenericButtonConfig } from "@/app/models/ButtonConfig";
-import { EventData, fetchEventData } from "@/app/services/eventsService";
+import { DynamicButtonProps, GenericButtonConfig } from "@/models/ButtonConfig";
+import { EventData, fetchEventData } from "@/services/eventsService";
 import { Component, useEffect, useState } from "react";
 import { Text, View, StyleSheet, ScrollView, ImageBackground } from "react-native";
-import DynamicButton from "../generic-buttons/dynamic-button";
+import DynamicButton from "../universal/dynamic-button";
 
 
 const DynamicEventSection = () => {
@@ -25,7 +25,7 @@ const DynamicEventSection = () => {
                 shape: 0, // full
                 buttonConfig: {
                     text: "Update Your Address",
-                    internalLink: `../sections/EventsPage/single-event-page`,
+                    internalLink: `../(tabs)/events`,
                     backgroundImage: "https://media.istockphoto.com/id/665336594/vector/blurred-summer-background-beach-with-sparkles-and-bokeh-vector-background-for-your-creativity.jpg?s=612x612&w=0&k=20&c=V7XOossa2nByJENpYqB-OpCONFOJS2oWI7j2Hkc8JIE="
                  }
             },
