+++ models/ButtonConfig.ts
